fixed import bugs on front end
