+++ frontend/src/Login.js
@@ -1,0 +1,8 @@
+import {Link} from "react-router-dom";
+import React from "react";
+
+function Login(){
+    return <h2> LOGIN </h2>
+}
+
+export default Login;
