+++ frontend/src/SignUp.js
@@ -1,0 +1,9 @@
+import {Link} from "react-router-dom";
+import React from "react";
+
+
+function SignUp() {
+    return <h2>SIGN UP</h2>;
+}
+
+export default SignUp;
