+++ frontend/src/index.js
@@ -3,7 +3,10 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import PageLayout from "./PageLayout";
-import {BrowserRouter, Route, Routes} from "react-router-dom";
+import Home from "./Home";
+import Login from "./Login";
+import SignUp from "./SignUp";
+import {BrowserRouter, Link, Route, Routes} from "react-router-dom";
 import reportWebVitals from './reportWebVitals';
